test: removed all unreliable locators
