+++ tests/addUserUI.spec.ts
@@ -47,9 +47,11 @@ test("Verify 'User Name' field design on the 'Add User' page", async ({
 test("Verify 'Year of Birth' field design and only number input on the 'Add User' page", async ({
   page,
 }) => {
-  const yearOfBirthField = page.getByPlaceholder("Year of Birth");
+  const yearOfBirthField = page.locator('xpath=//*[@id="inputYearOfBirth"]');
+  const placeholder = await yearOfBirthField.getAttribute("placeholder");
   await expect(yearOfBirthField).toBeVisible;
   await expect(yearOfBirthField).toHaveValue("");
+  await expect(placeholder).toEqual("Year of Birth");
 
   // check that non-number input is ignored be the Year of Birth field
   await yearOfBirthField.click();
@@ -62,16 +64,54 @@ test("Check 'Gender' field design and content on the 'Add User' page", async ({
 }) => {
   const genderField = page.locator('xpath=//*[@id="selectGender"]');
   await expect(genderField).toBeVisible;
-  // TODO: update locators which use Label
+
   // checking option 1 for gender input - Male
   await genderField.selectOption("1");
-  expect(await extractSelectedDisplayedValue(genderField)).toBe(GenderOptions[1]);
+  expect(await extractSelectedDisplayedValue(genderField)).toBe(
+    GenderOptions[1],
+  );
 
   // checking option 2 for gender input - Female
   await genderField.selectOption("2");
-  expect(await extractSelectedDisplayedValue(genderField)).toBe(GenderOptions[2]);
+  expect(await extractSelectedDisplayedValue(genderField)).toBe(
+    GenderOptions[2],
+  );
 
   // checking option 0 for gender input - Undefined
   await genderField.selectOption("0");
-  expect(await extractSelectedDisplayedValue(genderField)).toBe(GenderOptions[0]);
+  expect(await extractSelectedDisplayedValue(genderField)).toBe(
+    GenderOptions[0],
+  );
+});
+
+test("Verify Header design and content on the 'Add User' page", async ({
+  page,
+}) => {
+  // getting first listitem of the header
+  let listitem = page.locator("xpath=//ul/li[position()<2]");
+  await expect(listitem).toBeVisible;
+  await expect(listitem).toHaveRole("listitem");
+  // checking the content of the child of first listitem of the header, which is a link
+  let child = listitem.locator("xpath=child::*");
+  await expect(child).toHaveRole("link");
+  await expect(child).toContainText("Home");
+  await expect(child).toHaveAttribute("href", "/");
+
+  // getting the next listitem of the header
+  listitem = listitem.locator("xpath=following-sibling::li[1]");
+  await expect(listitem).toBeVisible();
+  await expect(listitem).toHaveRole("listitem");
+  // checking the content of the child of second listitem, which is a link
+  child = listitem.locator("xpath=descendant::*");
+  await expect(child).toHaveRole("link");
+  await expect(child).toContainText("Add User");
+  await expect(child).toHaveAttribute("href", "/Forms/AddUser");
+
+  // checking the content of last listitem of the header
+  child = listitem
+    .locator("xpath=ancestor::ul")
+    .locator('xpath=descendant::a[contains(text(),"Add Address")]');
+  await expect(child).toHaveRole("link");
+  await expect(child).toContainText("Add Address");
+  await expect(child).toHaveAttribute("href", "/Forms/AddAddress");
 });
